4_Fornecedor/consumer/requirements.txt
kafka-python==1.4.6
mysqlclient==1.4.2.post1
